+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.1.82] -- 2024-03-17
+
+- Add `normal_sub` to `test_sub` incident kinds.
+- Add `hashicorp_vault_mount`, `hashicorp_vault_mount`, `hashicorp_vault_mount` attributes to secret resource.
+- Add `incident_role_id` to `AddActionItemTaskParams`.
+
 ## [0.1.81] -- 2024-03-13
 
 - Add `service` to `workflow_task_invite_to_slack_channel` task.